Add Times class and its unit tests
A new Times class was added in the tabiya namespace along with corresponding unit tests. The Times class and its Iterator subclass are template classes that handle the restrictions of signed and unsigned integers. Times flows with the concepts defined in 'concept.hpp'. Also, CMakeLists files for both main project and tests were created. A '.gitignore' file is also added as part of the Git management tool.